+++ app/src/main/cpp/compat/vge.h
@@ -5,8 +5,50 @@
 extern "C" {
 #endif
 
+#define MAXSLOT 2
 #define XSIZE 240
 #define YSIZE 320
+
+struct _VRAM {
+    unsigned int pal[256];     /* パレット領域 */
+    unsigned char sp[0x20000]; /* スプライト領域 */
+};
+
+struct _SLOT {
+    int xs;
+    int ys;
+    unsigned char dat[65536];
+};
+
+struct _TOUCH {
+    int s;  /* 1:タッチ中, 0:リリース中 */
+    int dx; /* X方向移動ドット数 */
+    int dy; /* Y方向移動ドット数 */
+    int cx; /* 現在のX座標 */
+    int cy; /* 現在のY座標 */
+    int px; /* 直前のX座標 */
+    int py; /* 直前のY座標 */
+};
+
+struct Preferences {
+    int currentTitleId;
+    int base;
+    int infinity;
+    int loop;
+    int kobushi;
+    int localeId;
+};
+
+extern struct _VRAM _vram;
+extern struct _SLOT _slot[MAXSLOT];
+extern struct _TOUCH _touch;
+extern short *TONE1[85];
+extern short *TONE2[85];
+extern short *TONE3[85];
+extern short *TONE4[85];
+extern void *_psg;
+extern int g_flingY;
+extern int g_flingX;
 
 void tohovgs_cleanUp();
 
@@ -61,13 +103,15 @@ void vge_boxfSP(int fx, int fy, int tx, int ty, unsigned char c);
 
 void vge_touch(int *s, int *cx, int *cy, int *dx, int *dy);
 
-unsigned char vge_getmute();
-
 void vge_bplay(const char *mmlPath);
 
 void vge_bstop();
 
 void vge_bresume();
+
+void vge_restartCurrentSong();
+
+void vgsbuf(char *buf, size_t size);
 
 #ifdef __cplusplus
 };
